Code reduced
Had made a grave mistake and used readline() method instead of read() which means that i got the data in a list and not a string. Furthermore this meant that every bit of data had to be looped and stored in a new list and every first item in the list had to be changed and sewn back together.
TLDR: I had made it a hell for myself because I used the wrong function!